requirement.txt
altgraph==0.16.1
appdirs==1.4.3
asn1crypto==1.2.0
attrs==19.3.0
Automat==0.8.0
beautifulsoup4==4.8.1
bs4==0.0.1
captcha==0.3
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
dash==1.4.1
dash-core-components==1.3.1
dash-html-components==1.0.1
dash-renderer==1.1.2
dash-table==4.4.1
defusedxml==0.6.0
diff-match-patch==20121119
Django==2.0
django-captcha==1.0.0
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.1
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==2.0.13
django-simple-captcha==0.5.6
djangorestframework==3.10.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
Flask==1.1.1
Flask-Compress==1.4.0
future==0.16.0
httplib2==0.9.2
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
jdcal==1.4
Jinja2==2.10.3
lxml==4.4.1
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.3.6
olefile==0.45.1
openpyxl==2.5.3
parse==1.12.1
parsel==1.5.2
pefile==2019.4.18
Pillow==4.1.0
plotly==4.1.1
prettytable==0.7.2
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
PyDispatcher==2.0.5
pyecharts==1.5.1
pyee==6.0.0
pygame==1.9.6
PyHamcrest==1.9.0
PyInstaller==3.5
PyMySQL==0.8.1
pyOpenSSL==19.0.0
pyppeteer==0.0.25
pyquery==1.4.0
pytz==2018.4
pywin32-ctypes==0.2.0
PyYAML==3.12
queuelib==1.5.0
redis==3.3.11
requests==2.22.0
requests-html==0.10.0
retrying==1.3.3
Scrapy==1.7.3
scrapy-djangoitem==1.1.1
scrapyd==1.2.1
scrapyd-client==1.1.0
service-identity==18.1.0
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.4
SQLAlchemy==1.3.11
tablib==0.12.1
tqdm==4.36.1
Twisted==19.7.0
unicodecsv==0.14.1
urllib3==1.25.6
w3lib==1.21.0
websockets==8.0.2
Werkzeug==0.16.0
wincertstore==0.2
xlrd==1.1.0
xlwt==1.3.0
zope.interface==4.6.0
